requirements.txt
torch
torchvision
torchaudio
matplotlib
scikit-learn
pillow
numpy
pandas
streamlit
opencv-python-headless
